requirements.txt
torch==2.6.0
pytorch-lightning==2.5.5
transformers==4.57.1
PyYAML==6.0.2
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.6.1
torchmetrics==1.8.2
tokenizers==0.22.1
huggingface-hub==0.36.0
safetensors==0.6.2
